mをlong_long_intに

+++ src/spear.c
@@ -25,7 +25,7 @@ int main(){
     lb = 0;
     ub = woodSum / k + 1;
     while(ub - lb > 1){
-        int m = (lb + ub) / 2;
+        long long m = (lb + ub) / 2;
         if(p(m)){
             ub = m;
         }
